uapi: convert task_api.h
